+++ src/testrunner/main.cpp
@@ -27,12 +27,12 @@ int main(int argc, char **argv)
 #if defined(OS_WINDOWS)
 	{
 		char *selfEnd = strrchr(argv[0], '\\');
-		sprintf(testExe, "%.*s\\test.exe", selfEnd - argv[0], argv[0]);
+		sprintf(testExe, "%.*s\\test.exe", (long)(selfEnd - argv[0]), argv[0]);
 	}
 #elif defined(OS_LINUX)
 	{
 		char *selfEnd = strrchr(argv[0], '/');
-		sprintf(testExe, "%.*s/test", selfEnd - argv[0], argv[0]);
+		sprintf(testExe, "%.*s/test", (long)(selfEnd - argv[0]), argv[0]);
 	}
 #endif
